Johan, prøv å kjør dette på din maskin

+++ oblig1/main.c
@@ -3,7 +3,7 @@
 
 // These declarations are needed because we use .c files and not .h
 void read_graph_from_file(char *filename, int *N, int **row_ptr, int **col_idx, double **val);
-void PageRank_iterations(int N, int *row_ptr, int *col_idx, double *val, double d, double epsilon, double *scores);
+void PageRank_iterations(int N, int *row_ptr, int *col_idx, double *val, double d, long double epsilon, double *scores);
 void top_n_webpages(int N, double *scores, int n);
 
 
@@ -25,9 +25,11 @@ int main(int argc, char *argv[]){
     double *scores = (double*) malloc(N * sizeof(double));
 
     double d = atof(argv[2]);     // d in [0, 1], usually 0.85
-    double epsilon = atof(argv[3]);   // Convergence threshold
+    long double epsilon = (long double) atof(argv[3]);   // Convergence threshold
+    //printf("scores addr før: %p\n", scores);
 
     PageRank_iterations(N, row_ptr, col_idx, val, d, epsilon, scores);
+    //printf("scores addr etter: %p\n", scores);
 
     //printf("Addr scores: %p\n", scores);
     int n = atoi(argv[4]);
